+++ system/src/migrations/updates/_module.mjs
@@ -1,1 +1,2 @@
 export {default as Update_240218_1} from "./Update_240218_1.mjs";
+export {default as Update_240526_1} from "./Update_240526_1.mjs";